requirements.txt
appdirs==1.4.3
audioread==2.1.4
av==0.3.3
cycler==0.10.0
Cython==0.25.2
decorator==4.0.11
joblib==0.11
librosa==0.5.0
matplotlib==2.0.0
numpy==1.12.1
olefile==0.44
opencv-contrib-python==3.2.0.7
opencv-python==3.2.0.6
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
PySceneDetect==0.3.6
python-dateutil==2.6.0
pytz==2017.2
resampy==0.1.5
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
stft==0.5.2
termcolor==1.1.0
tesserocr==2.1.3
XlsxWriter==0.9.6
